+++ src/test/fixtures/simple-if-always-true/main.js
@@ -1,0 +1,9 @@
+function absWithBugs (x) {
+  if (x > 0) { // Should be <
+    x *= 1; // Should be -1
+  }
+  return x;
+}
+
+absWithBugs(3);
+absWithBugs(10);